requirements.txt
requests
discord.py[voice]
python-dotenv
opuslib
eyeD3
Pillow